Introduce encryption service
