+++ .env.example
@@ -22,3 +22,5 @@ GITHUB_CLIENT_ID=GitHub App client ID
 GITHUB_CLIENT_SECRET=GitHub App client secret
 GITHUB_APP_ID=123456
 GITHUB_PRIVATE_KEY_BASE_64=base 64 encoded version of the private key - see README.md for more info
+ENCRYPTION_PUBLIC_KEY_BASE_64=base 64 encoded version of the public key
+ENCRYPTION_PRIVATE_KEY_BASE_64=base 64 encoded version of the private key
